+++ src/main/java/dansplugins/wildpets/commands/HelpCommand.java
@@ -6,7 +6,7 @@ import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import dansplugins.wildpets.services.ConfigService;
+import dansplugins.wildpets.config.ConfigService;
 
 import java.util.ArrayList;
 import java.util.Arrays;
