+++ src/main/java/dansplugins/wildpets/commands/StatsCommand.java
@@ -1,6 +1,6 @@
 package dansplugins.wildpets.commands;
 
-import dansplugins.wildpets.data.PersistentData;
+import dansplugins.wildpets.pet.list.PetListRepository;
 import preponderous.ponder.minecraft.bukkit.abs.AbstractPluginCommand;
 
 import org.bukkit.ChatColor;
@@ -12,18 +12,18 @@ import java.util.Arrays;
  * @author Daniel McCoy Stephenson
  */
 public class StatsCommand extends AbstractPluginCommand {
-    private final PersistentData persistentData;
+    private final PetListRepository petListRepository;
 
     // TODO: add statistics object with persistent data like number of times players have attempted to tame
 
-    public StatsCommand(PersistentData persistentData) {
+    public StatsCommand(PetListRepository petListRepository) {
         super(new ArrayList<>(Arrays.asList("stats")), new ArrayList<>(Arrays.asList("wp.stats")));
-        this.persistentData = persistentData;
+        this.petListRepository = petListRepository;
     }
 
     public boolean execute(CommandSender sender) {
         sender.sendMessage(ChatColor.AQUA + "=== Wild Pets Statistics ===");
-        sender.sendMessage(ChatColor.AQUA + "Number of Pets: " + persistentData.getAllPets().size());
+        sender.sendMessage(ChatColor.AQUA + "Number of Pets: " + petListRepository.getAllPets().size());
         return true;
     }
