--- src/main/java/dansplugins/wildpets/objects/PetRecord.java
@@ -1,93 +1,0 @@
-package dansplugins.wildpets.objects;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
-import preponderous.ponder.misc.abs.Savable;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
-
-/**
- * @author Daniel McCoy Stephenson
- */
-public class PetRecord implements Savable {
-    private UUID uniqueID;
-    private String name;
-    private UUID ownerUUID;
-    private int assignedID;
-
-    public PetRecord(Pet pet) {
-        uniqueID = pet.getUniqueID();
-        name = pet.getName();
-        ownerUUID = pet.getOwnerUUID();
-        assignedID = pet.getAssignedID();
-    }
-
-    public PetRecord(Map<String, String> data) {
-        this.load(data);
-    }
-
-    public UUID getUniqueID() {
-        return uniqueID;
-    }
-
-    public void setUniqueID(UUID uniqueID) {
-        this.uniqueID = uniqueID;
-    }
-
-    public UUID getOwnerUUID() {
-        return ownerUUID;
-    }
-
-    public void setOwnerUUID(UUID ownerUUID) {
-        this.ownerUUID = ownerUUID;
-    }
-
-    public int getAssignedID() {
-        return assignedID;
-    }
-
-    public void setAssignedID(int assignedID) {
-        this.assignedID = assignedID;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public Map<String, String> save() {
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-
-        Map<String, String> saveMap = new HashMap<>();
-        saveMap.put("uniqueID", gson.toJson(uniqueID));
-        saveMap.put("owner", gson.toJson(ownerUUID));
-        saveMap.put("assignedID", gson.toJson(assignedID));
-        saveMap.put("name", gson.toJson(name));
-
-        return saveMap;
-    }
-
-    @Override
-    public void load(Map<String, String> data) {
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-
-        uniqueID = UUID.fromString(gson.fromJson(data.get("uniqueID"), String.class));
-        ownerUUID = UUID.fromString(gson.fromJson(data.get("owner"), String.class));
-        assignedID = Integer.parseInt(gson.fromJson(data.get("assignedID"), String.class));
-        name = gson.fromJson(data.get("name"), String.class);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (o == null || getClass() != o.getClass()) return false;
-        PetRecord petRecord = (PetRecord) o;
-        return uniqueID.equals(petRecord.uniqueID);
-    }
-}
